+++ TTS_xwing/src/Global.lua
@@ -116,7 +116,7 @@ StreamManagerModule = require("TTS_xwing.src.Player.StreamManager")
 --StreamManagerModule.AddPresentation("Blue", "right")
 
 local GlobalGuiControl = require("TTS_xwing.src.Player.GlobalGuiControl")
-local MoveData = require("TTS_xwing.src.MoveData")
+local MoveData = require("TTS_xwing.src.Game.Mechanic.Movement.MoveData")
 local AnnModule = require("TTS_xwing.src.Player.Announcements")
 local Maneuver = require("TTS_xwing.src.Game.Mechanic.Movement.Maneuver")
 local Relocator = Maneuver:new()
